Fix all warnings with gfortran 4.6.
Code can now be compiled safely with -Werror. :-)

+++ src/full_diagonalisation.F90
@@ -77,7 +77,7 @@ contains
 #endif
         end if
 
-        lwork = work(1)
+        lwork = nint(work(1))
         deallocate(work)
         call check_deallocate('work',ierr)
